editor/requirements.txt
etils[epath]
mlcroissant
numpy
pandas
pytest
python-magic
rdflib
requests
streamlit
streamlit-nested-layout
